+++ lib/compile.js
@@ -1,5 +1,3 @@
-var inlineRequire = require('../lib/inline-require.js');
-var serialize = require('../lib/serialize.js');
 var nearley = require('../lib/nearley.js');
 
 function Compile(structure, opts) {
@@ -10,10 +8,6 @@ function Compile(structure, opts) {
 
     function joiner(d) {
         return d.join('');
-    }
-
-    function id(d) {
-        return d[0];
     }
 
 	var outputRules = [];
@@ -64,46 +58,7 @@ function Compile(structure, opts) {
 		}
 	});
 
-    function variable(v, value) {
-        return "var " + assign(v, value);
-    }
-
-    function assign(v, value) {
-        return v + " = " + value + ";\n";
-    }
-
-    function comment(value) {
-        return "// " + value + "\n";
-    }
-
-    function wrap(body) {
-        return "function () {\n" + indent(body) + "}";
-    }
-
-    function indent(body) {
-        return body.replace(/^/mg, '    ');
-    }
-
-    function retn(stmt) {
-        return "return " + stmt;
-    }
-
-    function call(fn, params) {
-        return fn + "(" + params.join(', ') + ");";
-    }
-
-	var output = "";
-
-	output += comment("Generated automatically by nearley.");
-	output += assign(opts.export, wrap(
-        variable('nearley', inlineRequire(require, '../lib/nearley.js')) +
-        variable('id', serialize(id)) +
-        body.join('\n') +
-        "\n" +
-        retn(call("new nearley.Parser", [serialize(outputRules), JSON.stringify(firstName)]))
-    ));
-
-	return output;
+    return { rules: outputRules, body: body, start: firstName };
 }
 
 module.exports = Compile;
